downloader_api/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
cachetools==5.3.2
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.1
django-allauth
django-cors-headers==3.5.0
django-crispy-forms==1.11.0
django-currentuser
django-filter==2.4.0
django-rest-authtoken==2.1.3
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-user-agents==0.4.0
djangorestframework==3.14.0
dnspython==1.16.0
et-xmlfile==1.1.0
filestorage==0.0.6
google-api-core==2.13.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos

html5lib==1.1
httplib2==0.22.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
mysql-connector-python
mysqlclient==2.1.1
numpy==1.26.1
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.1.2
Pillow==10.0.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==3.17.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.3.0
reportlab==3.6.13
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.2
six==1.16.0
Solution==5.4.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ua-parser==0.18.0
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.2
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==3.0.1
xhtml2pdf==0.2.11
